% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_docker}
\alias{create_docker}
\title{Create a new docker object}
\usage{
create_docker(path = '')
}
\arguments{
\item{path}{character. A path to previously saved object. Keep empty for new object}
}
\value{
\code{docker} object.
}
\description{
\code{create_docker} is the prefered way to create a new docker object, either with default values, or by loading a previously saved configuration
}
\examples{
\dontrun{
# load existing config
docker = create_docker('old_config.RDS')

# create new
docker = create_docker()
}
}
\author{
Michal Kubista
}
