% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{status}
\alias{status}
\title{Post status}
\usage{
status(job, instance, metric, tags, value, host = "localhost:9091")
}
\arguments{
\item{job}{character. Top level name, such as project name.}

\item{instance}{character. Implementation name (project used for more client?)}

\item{metric}{character. Metric name, e.g. status, exec_time_s (execution time in seconds)}

\item{tags}{named vector(character=character). A vector of tag and tag values}

\item{value}{numeric. A value for the metric (status = 1, exec_time_s = 600)}

\item{value}{character. Hostname with port}
}
\value{
Returns an invisible \code{NULL} or character (in case of error)
}
\description{
\code{status} is a utility to post a value to Prometheus (pushgateway)
}
\examples{
\dontrun{
# successful job run
status("yieldigo", "penny", "status", c("run_at" = "14:00"), 1)

# job took 25 minutes
status("imdb", "update", "exec_time_s", c("scope" = "full", "delete_strategy"="truncate"), 25*60)

}
}
\author{
Michal Kubista
}
