% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{status_on_error}
\alias{status_on_error}
\title{Send status on error}
\usage{
status_on_error(job, instance, metric='error', tags, value, host = 'localhost:9091')
}
\arguments{
\item{job}{character. Top level name, such as project name.}

\item{instance}{character. Implementation name (project used for more client?)}

\item{metric}{character. Metric name, e.g. status, exec_time_s (execution time in seconds)}

\item{tags}{named vector(character=character). A vector of tag and tag values}

\item{value}{character. Hostname with port}
}
\value{
invisible \code{NULL}
}
\description{
\code{status_on_error} is a utility to post an error to Prometheus (pushgateway)
}
\examples{
\dontrun{
status_on_error('validation', 'client1', 'error', c('run_at' = '14:00'), 1)

}
}
\author{
Michal Kubista
}
